Add files via upload
I deployed the contract using meta mask, and inserted the address of the contract into the html file. The Abu was copied after the compilation in Remix. Nevertheless, whenever I push the button play (after opening the html), nothing happens. Button doesn't execute its function. By opening the developer tab and javascript console the problem seems to be associated with the impossibility to connect with provider, and a load failure persists whatever I change in the web 3 library versions, or the provider (local/given/current). I couldn't figure it out on how to connect the deployed smart contract and couldn't find the answers from the internet